Bound array indices, reject implausible clock writes, stop value truncation
Three unguarded array indices, all reachable without an attacker:

  webUIupdates.cpp indexed HC_PROGRAM[][9] with hc1_program/hc2_program
  and BURNER_STATE[][5] with kmStatus.BurnerStates - raw fields decoded
  straight from the serial bus. km271.cpp already clamps the same fields
  with limit() when it publishes them, but the web path did not, so one
  corrupted telegram (or a burner variant reporting an unexpected value)
  hands ArduinoJson a wild pointer. The bad value is latched into the
  status copy, so it would then repeat on every web UI page load.

  config.lang, config.mqtt.lang and config.log.filter are used as array
  indices too. The web callbacks validate them, but configLoadFromFile()
  does not, so a hand-edited or uploaded config.json could put an
  arbitrary value there. Clamped centrally in configFinalCheck(), which
  runs after every load, and cfg_logger_filter now clamps at the callback
  like cfg_lang already did.

km271SetDateTimeDTI()/km271SetDateTimeNTP() now refuse to write an
implausible date to the boiler. The automatic caller in main.cpp already
gates on getLocalTime(), but the cmd/datetime MQTT topic calls
km271SetDateTimeNTP() directly - so an HA automation firing right after a
reboot, before NTP resolved, could set the Logamatic's clock to 1970. The
web UI path builds its tm from separate date and time messages and could
submit a partially-filled struct. Since the Logamatic's day/night heating
program runs off this clock, a bogus value silently changes when the
house is heated.

km271sendCmd() took int8_t, so the narrowing happened before the range
check and the check validated the truncated value: publishing 286 to a
30..60 °C setpoint wrapped to 30 and was accepted as valid, with a
confirmation message reporting 30. Widened to int, so the existing range
checks see what was actually sent and reject it. The MQTT payload
variable was int16_t for the same reason - it also capped the oil counter
(in 1/100 L) at 327.67 L before wrapping negative - and is now long,
parsed with atol().

+++ include/km271.h
@@ -329,7 +329,7 @@ s_km271_alarm_str *km271GetAlarmMsgAdr();
 
 void sendKM271Info();
 void sendKM271Debug();
-void km271sendCmd(e_km271_sendCmd sendCmd, int8_t cmdPara);
+void km271sendCmd(e_km271_sendCmd sendCmd, int cmdPara);
 void km271sendCmdFlt(e_km271_sendCmd sendCmd, float cmdPara);
 bool km271sendServiceCmd(uint8_t cmdPara[8]);
 bool km271GetLogMode();
